backend/requirements.txt
Django
django-ninja
django-ninja-extra
django-cors-headers
